se migra proyecto a java 17

+++ src/main/java/com/digitalhouse/clinicaOdontologica/controller/OdontologoRestController.java
@@ -3,6 +3,7 @@ package com.digitalhouse.clinicaOdontologica.controller;
 
 import com.digitalhouse.clinicaOdontologica.domain.Odontologo;
 import com.digitalhouse.clinicaOdontologica.services.OdontologoService;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -82,14 +83,15 @@ public class OdontologoRestController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
-        try {
+    public ResponseEntity<String> delete(@PathVariable Long id) {
+        ResponseEntity<String> response = null;
+
+        if (odontologoService.findById(id).isPresent()) {
             odontologoService.delete(id);
-            logger.info("Se elimino el odontologo con id: " + id);
-            return ResponseEntity.ok().build();
-        } catch (ResponseStatusException e) {
-            logger.error("No se encontro el odontologo con id: " + id);
-            return ResponseEntity.status(e.getStatus()).build();
+            response = ResponseEntity.status(HttpStatus.NO_CONTENT).body("Eliminado");
+        } else {
+            response = ResponseEntity.status(HttpStatus.NOT_FOUND).build();
         }
+        return response;
     }
 }
